requirements.txt
#Python 3.10.8

py-cord == 2.2.2
xlrd == 1.2.0
python-dotenv == 0.21.0
Pillow == 9.3.0
Flask == 2.2.2
firebase-admin == 6.0.1
requests == 2.28.1